+++ workbench/old_solver/simulations_mp.py
@@ -1,23 +1,3 @@
-from pvlib import pvsystem, singlediode
-import numpy as np
-import multiprocess as mp
-from tqdm import notebook
-from workbench.host import module_mapping as ipv_mm
-from workbench.old_solver import calculations
-from workbench.host import host
-from workbench.utilities import circuits, general, temporal
-import time
-import copy
-import pandas as pd
-
-
-
-
-
-
-
-
-
 # def archive_mp_module_simulation(module_dict_chunk, module_name_chunk, cell_area, cell_params, timeseries):
 #     module_results = {}
 #
