+++ workbench/simulations/method_effective_irradiance.py
